Фикс твича

+++ twitchService/src/main/java/cersei/twitchservice/dto/GamesDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class GamesDto {
-    private Long gameId;
+    private Integer gameId;
     private String gameName;
 }
