+++ twitchService/src/main/java/cersei/twitchservice/model/Games.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Games {
     @Id
-    long gameId;
+    int gameId;
 
     String gameName;
 }
